+++ src/Exception/OutOfRangeException.php
@@ -1,6 +1,14 @@
 <?php
+/**
+ * SimpleComplex PHP Validate
+ * @link      https://github.com/simplecomplex/php-validate
+ * @copyright Copyright (c) 2020 Jacob Friis Mathiasen
+ * @license   https://github.com/simplecomplex/php-validate/blob/master/LICENSE (MIT License)
+ */
 
 namespace SimpleComplex\Validate\Exception;
+
+use SimpleComplex\Validate\Interfaces\ValidationExceptionInterface;
 
 /**
  * To differentiate exceptions thrown in-package from exceptions thrown
@@ -10,6 +18,6 @@ namespace SimpleComplex\Validate\Exception;
  *
  * @package SimpleComplex\Validate
  */
-class OutOfRangeException extends \OutOfRangeException
+class OutOfRangeException extends \OutOfRangeException implements ValidationExceptionInterface
 {
 }