+++ src/Exception/ValidationFailureException.php
@@ -1,6 +1,14 @@
 <?php
+/**
+ * SimpleComplex PHP Validate
+ * @link      https://github.com/simplecomplex/php-validate
+ * @copyright Copyright (c) 2020 Jacob Friis Mathiasen
+ * @license   https://github.com/simplecomplex/php-validate/blob/master/LICENSE (MIT License)
+ */
 
 namespace SimpleComplex\Validate\Exception;
+
+use SimpleComplex\Validate\Interfaces\ValidationExceptionInterface;
 
 /**
  * Generic validation failure exception - free to use within other packages.
@@ -10,7 +18,7 @@ namespace SimpleComplex\Validate\Exception;
  *
  * @package SimpleComplex\Validate
  */
-class ValidationFailureException extends \RuntimeException
+class ValidationFailureException extends \RuntimeException implements ValidationExceptionInterface
 {
     /**
      * @var string