+++ src/Interfaces/ValidationExceptionInterface.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * SimpleComplex PHP Validate
+ * @link      https://github.com/simplecomplex/php-validate
+ * @copyright Copyright (c) 2020 Jacob Friis Mathiasen
+ * @license   https://github.com/simplecomplex/php-validate/blob/master/LICENSE (MIT License)
+ */
+
+namespace SimpleComplex\Validate\Interfaces;
+
+interface ValidationExceptionInterface
+{
+}